# src/main/webapp/app/shared/api/generated/API-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "OncoKB™ is a precision oncology knowledge base developed at Memorial Sloan Kettering Cancer Center that contains biological and clinical information about genomic alterations in cancer.",
    "version": "v1.0.0",
    "title": "OncoKB™ APIs",
    "contact": {
      "name": "OncoKB™"
    },
    "license": {
      "name": "Terms of Use"
    }
  },
...
  "paths": {
...
    "/api/authenticate": {
      "get": {
        "tags": [
          "account-resource"
        ],
        "summary": "isAuthenticated",
        "operationId": "isAuthenticatedUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "user-uuid-controller"
        ],
        "summary": "authorize",
        "operationId": "authorizeUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "loginVM",
            "description": "loginVM",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginVM"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
    "/api/companies/{id}/users": {
      "get": {
        "tags": [
          "company-resource"
        ],
        "summary": "getCompanyUsers",
        "operationId": "getCompanyUsersUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/UserDTO"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
    "/api/mails/users": {
      "post": {
        "tags": [
          "mails-controller"
        ],
        "summary": "sendUserMails",
        "operationId": "sendUserMailsUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "by",
            "in": "query",
            "description": "by",
            "required": true,
            "type": "string"
          },
          {
            "name": "cc",
            "in": "query",
            "description": "cc",
            "required": false,
            "type": "string"
          },
          {
            "name": "from",
            "in": "query",
            "description": "from",
            "required": true,
            "type": "string"
          },
          {
            "name": "mailType",
            "in": "query",
            "description": "mailType",
            "required": true,
            "type": "string",
            "enum": [
              "ACTIVATION",
              "CREATION",
              "APPROVAL",
              "API_ACCESS_APPROVAL",
              "APPROVAL_ALIGN_LICENSE_WITH_COMPANY",
              "REJECTION",
              "REJECTION_US_SANCTION",
              "REJECT_ALUMNI_ADDRESS",
              "PASSWORD_RESET",
              "LICENSE_REVIEW_COMMERCIAL",
              "LICENSE_REVIEW_RESEARCH_COMMERCIAL",
              "LICENSE_REVIEW_HOSPITAL",
              "CLARIFY_ACADEMIC_FOR_PROFIT",
              "CLARIFY_ACADEMIC_NON_INSTITUTE_EMAIL",
              "CLARIFY_USE_CASE",
              "CLARIFY_DUPLICATE_USER",
              "CLARIFY_REGISTRATION_INFO",
              "LICENSE_OPTIONS",
              "VERIFY_EMAIL_BEFORE_ACCOUNT_EXPIRES",
              "ACTIVATE_FREE_TRIAL",
              "TRIAL_ACCOUNT_IS_ABOUT_TO_EXPIRE",
              "TRIAL_ACCOUNT_IS_ACTIVATED",
              "DATA_USAGE_EXCEEDS_THRESHOLD",
              "TOKEN_HAS_BEEN_EXPOSED",
              "TOKEN_HAS_BEEN_EXPOSED_USER",
              "SEARCHING_RESPONSE_STRUCTURE_HAS_CHANGED",
              "LIST_OF_UNAPPROVED_USERS",
              "TERMINATION_NOTIFICATION_EMAIL",
              "TEST"
            ]
          },
          {
            "name": "to",
            "in": "query",
            "description": "to",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
    "/api/tokens/{uuid}": {
      "get": {
        "tags": [
          "token-resource"
        ],
        "summary": "getToken",
        "operationId": "getTokenUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "description": "uuid",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Token"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "token-resource"
        ],
        "summary": "deleteTokenWithoutTransferringStats",
        "operationId": "deleteTokenWithoutTransferringStatsUsingDELETE",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "description": "uuid",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
...
    "/api/usage/summary/users": {
      "get": {
        "tags": [
          "usage-analysis-controller"
        ],
        "summary": "userOverviewUsageGet",
        "operationId": "userOverviewUsageGetUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "companyId",
            "in": "query",
            "description": "companyId",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/UserOverviewUsage"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/usage/users/{userId}": {
      "get": {
        "tags": [
          "usage-analysis-controller"
        ],
        "summary": "userUsageGet",
        "operationId": "userUsageGetUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "userId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserUsage"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
    "/api/user-mails/users/{login}": {
      "get": {
        "tags": [
          "user-mails-resource"
        ],
        "summary": "getUsersUserMails",
        "operationId": "getUsersUserMailsUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "description": "login",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/UserMailsDTO"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
    "/api/users": {
      "get": {
        "tags": [
          "user-resource"
        ],
        "summary": "getAllUsers",
        "operationId": "getAllUsersUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/UserDTO"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "user-resource"
        ],
        "summary": "createUser",
        "operationId": "createUserUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "managedUserVM",
            "description": "managedUserVM",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ManagedUserVM"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": [
          "user-resource"
        ],
        "summary": "updateUser",
        "operationId": "updateUserUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "sendEmail",
            "in": "query",
            "description": "sendEmail",
            "required": true,
            "type": "boolean"
          },
          {
            "name": "unlinkUser",
            "in": "query",
            "description": "unlinkUser",
            "required": true,
            "type": "boolean"
          },
          {
            "in": "body",
            "name": "userDTO",
            "description": "userDTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserDTO"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
    "/api/users/{login}": {
      "get": {
        "tags": [
          "user-resource"
        ],
        "summary": "getUser",
        "operationId": "getUserUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "description": "login",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserDTO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "user-resource"
        ],
        "summary": "deleteUser",
        "operationId": "deleteUserUsingDELETE",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "description": "login",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
...
    "AdditionalInfoDTO": {
      "type": "object",
      "properties": {
        "apiAccessRequest": {
          "$ref": "#/definitions/ApiAccessRequest"
        },
        "trialAccount": {
          "$ref": "#/definitions/TrialAccount"
        },
        "userCompany": {
          "$ref": "#/definitions/UserCompany"
        }
      },
      "title": "AdditionalInfoDTO"
    },
    "ApiAccessRequest": {
      "type": "object",
      "properties": {
        "justification": {
          "type": "string"
        },
        "requested": {
          "type": "boolean"
        }
      },
      "title": "ApiAccessRequest"
    },
    "CompanyAdditionalInfoDTO": {
      "type": "object",
      "properties": {
        "license": {
          "$ref": "#/definitions/CompanyLicense"
        }
      },
      "title": "CompanyAdditionalInfoDTO"
    },
    "CompanyDTO": {
      "type": "object",
      "required": [
        "companyType",
        "licenseModel",
        "licenseStatus",
        "licenseType",
        "name"
      ],
      "properties": {
        "additionalInfo": {
          "$ref": "#/definitions/CompanyAdditionalInfoDTO"
        },
        "businessContact": {
          "type": "string"
        },
        "companyDomains": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "companyType": {
          "type": "string",
          "enum": [
            "PARENT",
            "BRANCH",
            "UNKNOWN"
          ]
        },
        "description": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "legalContact": {
          "type": "string"
        },
        "licenseModel": {
          "type": "string",
          "enum": [
            "FULL",
            "LIMITED"
          ]
        },
        "licenseStatus": {
          "type": "string",
          "enum": [
            "TRIAL",
            "REGULAR",
            "TRIAL_EXPIRED",
            "EXPIRED",
            "UNKNOWN"
          ]
        },
        "licenseType": {
          "type": "string",
          "enum": [
            "ACADEMIC",
            "COMMERCIAL",
            "RESEARCH_IN_COMMERCIAL",
            "HOSPITAL"
          ]
        },
        "name": {
          "type": "string"
        },
        "numberOfUsers": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "CompanyDTO"
    },
...
    "LoginVM": {
      "type": "object",
      "required": [
        "password",
        "username"
      ],
      "properties": {
        "password": {
          "type": "string",
          "minLength": 4,
          "maxLength": 100
        },
        "rememberMe": {
          "type": "boolean"
        },
        "username": {
          "type": "string",
          "minLength": 1,
          "maxLength": 50
        }
      },
      "title": "LoginVM"
    },
...
    "ManagedUserVM": {
      "type": "object",
      "properties": {
        "activated": {
          "type": "boolean"
        },
        "activationKey": {
          "type": "string"
        },
        "additionalInfo": {
          "$ref": "#/definitions/AdditionalInfoDTO"
        },
        "authorities": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "city": {
          "type": "string",
          "minLength": 2,
          "maxLength": 2147483647
        },
        "company": {
          "$ref": "#/definitions/CompanyDTO"
        },
        "companyName": {
          "type": "string",
          "minLength": 2,
          "maxLength": 2147483647
        },
        "country": {
          "type": "string",
          "minLength": 2,
          "maxLength": 2147483647
        },
        "createdBy": {
          "type": "string"
        },
        "createdDate": {
          "type": "string",
          "format": "date-time"
        },
        "email": {
          "type": "string",
          "minLength": 5,
          "maxLength": 254
        },
        "emailVerified": {
          "type": "boolean"
        },
        "firstName": {
          "type": "string",
          "minLength": 2,
          "maxLength": 50
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "imageUrl": {
          "type": "string",
          "minLength": 0,
          "maxLength": 256
        },
        "jobTitle": {
          "type": "string"
        },
        "langKey": {
          "type": "string",
          "minLength": 2,
          "maxLength": 10
        },
        "lastModifiedBy": {
          "type": "string"
        },
        "lastModifiedDate": {
          "type": "string",
          "format": "date-time"
        },
        "lastName": {
          "type": "string",
          "minLength": 2,
          "maxLength": 50
        },
        "licenseType": {
          "type": "string",
          "enum": [
            "ACADEMIC",
            "COMMERCIAL",
            "RESEARCH_IN_COMMERCIAL",
            "HOSPITAL"
          ]
        },
        "login": {
          "type": "string",
          "minLength": 1,
          "maxLength": 50,
          "pattern": "^[_.\\-\\+@A-Za-z0-9]*$"
        },
        "notifyUserOnTrialCreation": {
          "type": "boolean"
        },
        "password": {
          "type": "string",
          "minLength": 4,
          "maxLength": 100
        },
        "resetDate": {
          "type": "string",
          "format": "date-time"
        },
        "resetKey": {
          "type": "string"
        },
        "tokenIsRenewable": {
          "type": "boolean"
        },
        "tokenValidDays": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "ManagedUserVM"
    },
...
    "Token": {
      "type": "object",
      "required": [
        "currentUsage",
        "renewable"
      ],
      "properties": {
        "creation": {
          "type": "string",
          "format": "date-time"
        },
        "currentUsage": {
          "type": "integer",
          "format": "int32"
        },
        "expiration": {
          "type": "string",
          "format": "date-time"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "renewable": {
          "type": "boolean"
        },
        "token": {
          "type": "string"
        },
        "usageLimit": {
          "type": "integer",
          "format": "int32"
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      },
      "title": "Token"
    },
    "TrialAccount": {
      "type": "object",
      "properties": {
        "activation": {
          "$ref": "#/definitions/Activation"
        },
        "licenseAgreement": {
          "$ref": "#/definitions/LicenseAgreement"
        }
      },
      "title": "TrialAccount"
    },
    "UsageSummary": {
      "type": "object",
      "properties": {
        "day": {
          "type": "object",
          "additionalProperties": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int64"
            }
          }
        },
        "month": {
          "type": "object",
          "additionalProperties": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int64"
            }
          }
        },
        "year": {
          "type": "object",
          "additionalProperties": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
      "title": "UsageSummary"
    },
    "User": {
      "type": "object",
      "required": [
        "activated",
        "login"
      ],
      "properties": {
        "activated": {
          "type": "boolean"
        },
        "email": {
          "type": "string",
          "minLength": 5,
          "maxLength": 254
        },
        "firstName": {
          "type": "string",
          "minLength": 0,
          "maxLength": 50
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "imageUrl": {
          "type": "string",
          "minLength": 0,
          "maxLength": 256
        },
        "langKey": {
          "type": "string",
          "minLength": 2,
          "maxLength": 10
        },
        "lastName": {
          "type": "string",
          "minLength": 0,
          "maxLength": 50
        },
        "login": {
          "type": "string",
          "minLength": 1,
          "maxLength": 50,
          "pattern": "^[_.\\-\\+@A-Za-z0-9]*$"
        },
        "resetDate": {
          "type": "string",
          "format": "date-time"
        }
      },
      "title": "User"
    },
...
    "UserDTO": {
      "type": "object",
      "properties": {
        "activated": {
          "type": "boolean"
        },
        "activationKey": {
          "type": "string"
        },
        "additionalInfo": {
          "$ref": "#/definitions/AdditionalInfoDTO"
        },
        "authorities": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "city": {
          "type": "string",
          "minLength": 2,
          "maxLength": 2147483647
        },
        "company": {
          "$ref": "#/definitions/CompanyDTO"
        },
        "companyName": {
          "type": "string",
          "minLength": 2,
          "maxLength": 2147483647
        },
        "country": {
          "type": "string",
          "minLength": 2,
          "maxLength": 2147483647
        },
        "createdBy": {
          "type": "string"
        },
        "createdDate": {
          "type": "string",
          "format": "date-time"
        },
        "email": {
          "type": "string",
          "minLength": 5,
          "maxLength": 254
        },
        "emailVerified": {
          "type": "boolean"
        },
        "firstName": {
          "type": "string",
          "minLength": 2,
          "maxLength": 50
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "imageUrl": {
          "type": "string",
          "minLength": 0,
          "maxLength": 256
        },
        "jobTitle": {
          "type": "string"
        },
        "langKey": {
          "type": "string",
          "minLength": 2,
          "maxLength": 10
        },
        "lastModifiedBy": {
          "type": "string"
        },
        "lastModifiedDate": {
          "type": "string",
          "format": "date-time"
        },
        "lastName": {
          "type": "string",
          "minLength": 2,
          "maxLength": 50
        },
        "licenseType": {
          "type": "string",
          "enum": [
            "ACADEMIC",
            "COMMERCIAL",
            "RESEARCH_IN_COMMERCIAL",
            "HOSPITAL"
          ]
        },
        "login": {
          "type": "string",
          "minLength": 1,
          "maxLength": 50,
          "pattern": "^[_.\\-\\+@A-Za-z0-9]*$"
        },
        "resetDate": {
          "type": "string",
          "format": "date-time"
        },
        "resetKey": {
          "type": "string"
        }
      },
      "title": "UserDTO"
    },
...
    "UserMailsDTO": {
      "type": "object",
      "required": [
        "sentBy",
        "sentDate",
        "sentFrom"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "mailType": {
          "type": "string",
          "enum": [
            "ACTIVATION",
            "CREATION",
            "APPROVAL",
            "API_ACCESS_APPROVAL",
            "APPROVAL_ALIGN_LICENSE_WITH_COMPANY",
            "REJECTION",
            "REJECTION_US_SANCTION",
            "REJECT_ALUMNI_ADDRESS",
            "PASSWORD_RESET",
            "LICENSE_REVIEW_COMMERCIAL",
            "LICENSE_REVIEW_RESEARCH_COMMERCIAL",
            "LICENSE_REVIEW_HOSPITAL",
            "CLARIFY_ACADEMIC_FOR_PROFIT",
            "CLARIFY_ACADEMIC_NON_INSTITUTE_EMAIL",
            "CLARIFY_USE_CASE",
            "CLARIFY_DUPLICATE_USER",
            "CLARIFY_REGISTRATION_INFO",
            "LICENSE_OPTIONS",
            "VERIFY_EMAIL_BEFORE_ACCOUNT_EXPIRES",
            "ACTIVATE_FREE_TRIAL",
            "TRIAL_ACCOUNT_IS_ABOUT_TO_EXPIRE",
            "TRIAL_ACCOUNT_IS_ACTIVATED",
            "DATA_USAGE_EXCEEDS_THRESHOLD",
            "TOKEN_HAS_BEEN_EXPOSED",
            "TOKEN_HAS_BEEN_EXPOSED_USER",
            "SEARCHING_RESPONSE_STRUCTURE_HAS_CHANGED",
            "LIST_OF_UNAPPROVED_USERS",
            "TERMINATION_NOTIFICATION_EMAIL",
            "TEST"
          ]
        },
        "sentBy": {
          "type": "string"
        },
        "sentDate": {
          "type": "string",
          "format": "date-time"
        },
        "sentFrom": {
          "type": "string"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        },
        "userLogin": {
          "type": "string"
        }
      },
      "title": "UserMailsDTO"
    },
    "UserOverviewUsage": {
      "type": "object",
      "properties": {
        "dayUsage": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/UserStats"
          }
        },
        "monthUsage": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/UserStats"
          }
        },
        "userEmail": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "yearUsage": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/UserStats"
          }
        }
      },
      "title": "UserOverviewUsage"
    },
    "UserStats": {
      "type": "object",
      "properties": {
        "maxUsageProportion": {
          "type": "number",
          "format": "float"
        },
        "mostUsedEndpoint": {
          "type": "string"
        },
        "mostUsedPublicEndpoint": {
          "type": "string"
        },
        "publicMaxUsageProportion": {
          "type": "number",
          "format": "float"
        },
        "totalPublicUsage": {
          "type": "integer",
          "format": "int64"
        },
        "totalUsage": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "UserStats"
    },
    "UserUsage": {
      "type": "object",
      "properties": {
        "company": {
          "type": "string"
        },
        "jobTitle": {
          "type": "string"
        },
        "licenseType": {
          "type": "string"
        },
        "summary": {
          "$ref": "#/definitions/UsageSummary"
        },
        "userEmail": {
          "type": "string"
        },
        "userFirstName": {
          "type": "string"
        },
        "userLastName": {
          "type": "string"
        }
      },
      "title": "UserUsage"
    },
...
  }
}